requirements.txt
fastapi[standard]
python-socketio
paramiko
python-dotenv